refactor(mb-api): extraire les loaders DB partagés (sous-arbre, agrégation, individus)
loadSousArbre, loadFonctionsAgregation et loadIndividusParPublicId étaient dupliqués
entre valeurAvancement et objectifIndicateurIndividu. Regroupés dans
indicateur/queries/loadIndicateurIndividuContext.ts.
