+++ apps/mb-api/src/valeurAvancement/queries/loadResolveSerieContext.ts
@@ -1,8 +1,11 @@
-import { type FonctionAgregation } from '@pilote/mb-shared/indicateur'
 import { type DateTrunc } from '@pilote/mb-shared/valeurAvancement'
 
 import { db } from '@/framework/persistence/dbStore'
 import { getValeursTronqueesPourIndividus } from '@/generated/prisma/sql'
+import {
+  loadFonctionsAgregation,
+  loadSousArbre,
+} from '@/indicateur/queries/loadIndicateurIndividuContext'
 import {
   type IndividuRef,
   type ResolveSerieContext,
@@ -43,58 +46,6 @@ export const loadResolveSerieContext = async ({
     ),
   }
   return { ctx, allNodes }
-}
-
-// BFS itératif niveau par niveau plutôt qu'une CTE récursive PostgreSQL :
-// l'arbre est typiquement peu profond (3-4 niveaux France→Région→Département)
-// et garder le calcul en JS reste plus simple à lire et à débugger que du SQL
-// récursif. Détection naïve des cycles : on ne retraite pas un nœud déjà vu.
-const loadSousArbre = async (
-  cibles: ReadonlyArray<IndividuRef>,
-): Promise<{
-  allNodes: Map<string, IndividuRef>
-  enfantsParParent: Map<string, IndividuRef[]>
-}> => {
-  const allNodes = new Map<string, IndividuRef>()
-  for (const cible of cibles) allNodes.set(cible.id, cible)
-  const enfantsParParent = new Map<string, IndividuRef[]>()
-  let currentLevel = cibles.map((c) => c.id)
-
-  while (currentLevel.length > 0) {
-    const relations = await db().relation.findMany({
-      where: { parentId: { in: currentLevel } },
-      include: { child: true },
-    })
-    if (relations.length === 0) break
-
-    const nextLevelSet = new Set<string>()
-    for (const relation of relations) {
-      const childRef: IndividuRef = {
-        id: relation.child.id,
-        publicId: relation.child.publicId,
-        referentielId: relation.child.referentielId,
-      }
-      if (!allNodes.has(childRef.id)) {
-        allNodes.set(childRef.id, childRef)
-        nextLevelSet.add(childRef.id)
-      }
-      const liste = enfantsParParent.get(relation.parentId) ?? []
-      liste.push(childRef)
-      enfantsParParent.set(relation.parentId, liste)
-    }
-    currentLevel = [...nextLevelSet]
-  }
-
-  return { allNodes, enfantsParParent }
-}
-
-const loadFonctionsAgregation = async (
-  indicateurId: string,
-): Promise<Map<string, FonctionAgregation>> => {
-  const liens = await db().indicateurReferentiel.findMany({
-    where: { indicateurId },
-  })
-  return new Map(liens.map((lien) => [lien.referentielId, lien.fonctionAgregation]))
 }
 
 const loadSaisiesTronquees = async ({
